requirements.txt
flask_sqlalchemy=2.0.43
flask_login=0.6.3
fatsecret-py=0.4.0
rauth=0.7.3